Rename meshimporter as modelimporter
