+++ Engine/Main/Application.cpp
@@ -16,7 +16,7 @@
 #include "Module/ModuleFileSystem.h"
 #include "Module/ModuleWindow.h"
 #include "ResourceManagement/Importer/MaterialImporter.h"
-#include "ResourceManagement/Importer/MeshImporter.h"
+#include "ResourceManagement/Importer/ModelImporter.h"
 #include "UI/EngineLog.h"
 #include "Brofiler/Brofiler.h"
